Consolidating store & reducers into store.ts
