+++ src/app/modules/timeline/selectors.ts
@@ -1,4 +1,4 @@
-import { RootState } from '../../store/reducers';
+import { RootState } from '../../store';
 
 export const visibleTime = (state: RootState): { visibleTimeStart: any, visibleTimeEnd: any } => {
     const {visibleTimeStart,visibleTimeEnd} = state.timeline;
